% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R, R/zzz-aliases.R
\name{bulk_project_metadata}
\alias{bulk_project_metadata}
\alias{BulkProjectMetadata}
\title{Bulk project metadata}
\usage{
bulk_project_metadata(con, accessions)

BulkProjectMetadata(con, accessions)
}
\arguments{
\item{con}{A \code{seqout_connection}.}

\item{accessions}{Character vector of project accessions.}
}
\value{
A tibble with project metadata.
}
\description{
Retrieve accession, title, and description for multiple projects.
}
