% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{close_seqout}
\alias{close_seqout}
\title{Close a SeqOut connection}
\usage{
close_seqout(con)
}
\arguments{
\item{con}{A \code{seqout_connection} returned by \code{\link[=connect_seqout]{connect_seqout()}}.}
}
\description{
Close a SeqOut connection
}
