% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organisms.R, R/zzz-aliases.R
\name{common_name}
\alias{common_name}
\alias{CommonName}
\title{Common name for a scientific name}
\usage{
common_name(scientific_name, con = .con())

CommonName(scientific_name, con = .con())
}
\arguments{
\item{scientific_name}{A scientific name, for example \code{"Homo sapiens"}.}

\item{con}{A \code{seqout_connection}. Defaults to the shared REST connection.}
}
\value{
A tibble with \code{tax_id}, \code{scientific_name} and \code{common_name}.
}
\description{
Common name for a scientific name
}
\examples{
\dontrun{
common_name("Homo sapiens")
}
}
