% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R, R/zzz-aliases.R
\name{counts_matrix}
\alias{counts_matrix}
\alias{genexcell_counts}
\alias{cellxgene_counts}
\alias{CountsMatrix}
\title{Counts in either orientation}
\usage{
counts_matrix(x)

genexcell_counts(x)

cellxgene_counts(x)

CountsMatrix(x)
}
\arguments{
\item{x}{A \code{seqout_matrix} from \code{\link[=seqout_matrix]{seqout_matrix()}} or \code{\link[=matrices]{matrices()}}.}
}
\value{
A matrix, dgCMatrix if \code{x$X} was sparse.
}
\description{
\code{counts_matrix()}, aliased \code{genexcell_counts()}, gives features by
observations; \code{cellxgene_counts()} the reverse.
}
\examples{
\dontrun{
counts <- seqout_counts(gse = "GSE291735")
mats <- lapply(wt$sample, function(gsm) counts_matrix(seqout_matrix(counts, sample = gsm)))
}

}
