% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R, R/zzz-aliases.R
\name{download_script}
\alias{download_script}
\alias{DownloadScript}
\title{Get a download script for a study}
\usage{
download_script(study, mode = "fastq", file = NULL, con = .con())

DownloadScript(study, mode = "fastq", file = NULL, con = .con())
}
\arguments{
\item{study}{A study accession.}

\item{mode}{Download source.}

\item{file}{Optional path to write the script to.}

\item{con}{A \code{seqout_connection}. Defaults to the shared REST connection.}
}
\value{
The script as a character string, invisibly when \code{file} is given.
}
\description{
Get a download script for a study
}
