% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R, R/zzz-aliases.R
\name{experiment_runs}
\alias{experiment_runs}
\alias{ExperimentRuns}
\title{The runs of one experiment}
\usage{
experiment_runs(con, accession)

ExperimentRuns(con, accession)
}
\arguments{
\item{con}{A \code{seqout_connection} from \code{\link[=seqout_connect]{seqout_connect()}}.}

\item{accession}{An experiment accession (SRX, ERX, DRX, CRX, HRX).}
}
\value{
A tibble of runs.
}
\description{
The runs of one experiment
}
