% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R, R/zzz-aliases.R
\name{global_contributions}
\alias{global_contributions}
\alias{GlobalContributions}
\title{Get global geographic contributions}
\usage{
global_contributions(
  organism = NULL,
  assay_l1 = NULL,
  assay_l2 = NULL,
  place_type = NULL,
  address_type = NULL,
  con = .con()
)

GlobalContributions(
  organism = NULL,
  assay_l1 = NULL,
  assay_l2 = NULL,
  place_type = NULL,
  address_type = NULL,
  con = .con()
)
}
\arguments{
\item{organism}{Optional. Filter by scientific name.}

\item{assay_l1}{Optional. Filter by assay level 1.}

\item{assay_l2}{Optional. Filter by assay level 2.}

\item{place_type}{Optional. Filter by place type.}

\item{address_type}{Optional. Filter by address type.}

\item{con}{A \code{seqout_connection}. Defaults to the shared REST connection.}
}
\value{
A tibble with geographic contribution data.
}
\description{
Get global geographic contributions
}
