% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R, R/zzz-aliases.R
\name{paper}
\alias{paper}
\alias{Paper}
\title{Find projects linked to a publication}
\usage{
paper(id = NULL, pmid = NULL, doi = NULL, con = .con())

Paper(id = NULL, pmid = NULL, doi = NULL, con = .con())
}
\arguments{
\item{id}{Bare PubMed ID (digits) or DOI starting \code{10.}. URLs error.}

\item{pmid}{PubMed ID.}

\item{doi}{DOI.}

\item{con}{A \code{seqout_connection}. Defaults to the shared REST connection.}
}
\value{
A tibble of linked projects, empty when the publication is unknown.
}
\description{
Find projects linked to a publication
}
\examples{
\dontshow{if (SeqoutOnline()) withAutoprint(\{ # examplesIf}
Paper("34764296")
Paper(doi = "10.1038/s41467-021-26864-x")
\dontshow{\}) # examplesIf}
}
