% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R, R/zzz-aliases.R
\name{seqout_sample}
\alias{seqout_sample}
\alias{SeqoutSample}
\title{Get SRA/ENA sample metadata}
\usage{
seqout_sample(con, accession)

SeqoutSample(con, accession)
}
\arguments{
\item{con}{A \code{seqout_connection}.}

\item{accession}{Sample accession (SRS/DRS/ERS/SAM*/GSM*).}
}
\value{
A tibble with sample metadata.
}
\description{
Get SRA/ENA sample metadata
}
