% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex.R, R/zzz-aliases.R
\name{seqout_sex_plot}
\alias{seqout_sex_plot}
\alias{SeqoutSexPlot}
\title{Plot dual sex-chromosome evidence}
\usage{
seqout_sex_plot(x)

SeqoutSexPlot(x)
}
\arguments{
\item{x}{A result from \code{\link[=seqout_sex]{seqout_sex()}}.}
}
\value{
A ggplot object.
}
\description{
Plot dual sex-chromosome evidence
}
