% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R
\name{sq_bulk_project_metadata}
\alias{sq_bulk_project_metadata}
\title{Bulk project metadata}
\usage{
sq_bulk_project_metadata(con, accessions)
}
\arguments{
\item{con}{A \code{seqout_connection}.}

\item{accessions}{Character vector of project accessions.}
}
\value{
A tibble with project metadata.
}
\description{
Retrieve accession, title, and description for multiple projects.
}
