% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{sq_clear_cache}
\alias{sq_clear_cache}
\title{Clear locally cached tables}
\usage{
sq_clear_cache(con)
}
\arguments{
\item{con}{A \code{seqout_connection}.}
}
\value{
Number of tables removed (invisibly).
}
\description{
Removes all \verb{*_local} tables created by \code{\link[=sq_cache_table]{sq_cache_table()}}.
}
