% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{sq_download_script}
\alias{sq_download_script}
\title{Get a download script for a study}
\usage{
sq_download_script(con, study, mode = "fastq", file = NULL)
}
\arguments{
\item{con}{A \code{seqout_connection}.}

\item{study}{Character. Study accession.}

\item{mode}{Download source: \code{"fastq"}, \code{"sra"}, \code{"sra_lite"}, \code{"s3"},
or \code{"gcs"}.}

\item{file}{Optional file path to write the script to.}
}
\value{
The script as a character string (invisibly if \code{file} is provided).
}
\description{
Get a download script for a study
}
