% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{sq_fastq_links}
\alias{sq_fastq_links}
\title{Get FASTQ download links for a study}
\usage{
sq_fastq_links(con, study)
}
\arguments{
\item{con}{A \code{seqout_connection}.}

\item{study}{Character. Study accession.}
}
\value{
A tibble with run accessions and FASTQ download URLs.
}
\description{
Get FASTQ download links for a study
}
\seealso{
\code{\link[=sq_project_runs]{sq_project_runs()}}
}
