% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sq_global_contributions}
\alias{sq_global_contributions}
\title{Get global geographic contributions}
\usage{
sq_global_contributions(
  con,
  organism = NULL,
  assay_l1 = NULL,
  assay_l2 = NULL,
  place_type = NULL,
  address_type = NULL
)
}
\arguments{
\item{con}{A \code{seqout_connection}.}

\item{organism}{Optional. Filter by scientific name.}

\item{assay_l1}{Optional. Filter by assay level 1.}

\item{assay_l2}{Optional. Filter by assay level 2.}

\item{place_type}{Optional. Filter by place type.}

\item{address_type}{Optional. Filter by address type.}
}
\value{
A tibble with geographic contribution data.
}
\description{
Get global geographic contributions
}
