% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{sq_project_experiments}
\alias{sq_project_experiments}
\title{Get experiments for a study}
\usage{
sq_project_experiments(con, study)
}
\arguments{
\item{con}{A \code{seqout_connection}.}

\item{study}{Character. Study accession (SRP/ERP/DRP/PRJ*).}
}
\value{
A tibble of experiment metadata.
}
\description{
Get experiments for a study
}
