% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{sq_project_runs}
\alias{sq_project_runs}
\title{Get run download links for a study}
\usage{
sq_project_runs(con, study)
}
\arguments{
\item{con}{A \code{seqout_connection}.}

\item{study}{Character. Study accession.}
}
\value{
A tibble of run metadata with download links.
}
\description{
Get run download links for a study
}
