% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{sq_sample}
\alias{sq_sample}
\title{Get SRA/ENA sample metadata}
\usage{
sq_sample(con, accession)
}
\arguments{
\item{con}{A \code{seqout_connection}.}

\item{accession}{Sample accession (SRS/DRS/ERS/SAM*/GSM*).}
}
\value{
A tibble with sample metadata.
}
\description{
Get SRA/ENA sample metadata
}
