% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk.R, R/zzz-aliases.R
\name{bulk_project_metadata}
\alias{bulk_project_metadata}
\alias{BulkProjectMetadata}
\title{Bulk project metadata}
\usage{
bulk_project_metadata(accessions, con = .con())

BulkProjectMetadata(accessions, con = .con())
}
\arguments{
\item{accessions}{Character vector of project accessions.}

\item{con}{A \code{seqout_connection}. Defaults to the shared REST connection.}
}
\value{
A tibble with project metadata.
}
\description{
Retrieve accession, title, and description for multiple projects.
}
