% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R, R/zzz-aliases.R
\name{clear_cache}
\alias{clear_cache}
\alias{ClearCache}
\title{Clear locally cached tables}
\usage{
clear_cache(con = .con())

ClearCache(con = .con())
}
\arguments{
\item{con}{A Parquet \code{seqout_connection} from \code{\link[=seqout_connect]{seqout_connect()}}.}
}
\value{
Number of tables removed (invisibly).
}
\description{
Removes all \verb{*_local} tables created by \code{\link[=cache_table]{cache_table()}}.
}
