% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R, R/zzz-aliases.R
\name{clear_cache}
\alias{clear_cache}
\alias{ClearCache}
\title{Clear locally cached tables}
\usage{
clear_cache(con)

ClearCache(con)
}
\arguments{
\item{con}{A \code{seqout_connection}.}
}
\value{
Number of tables removed (invisibly).
}
\description{
Removes all \verb{*_local} tables created by \code{\link[=cache_table]{cache_table()}}.
}
