% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organisms.R, R/zzz-aliases.R
\name{common_name}
\alias{common_name}
\alias{CommonName}
\title{Common name for a scientific name}
\usage{
common_name(con, scientific_name)

CommonName(con, scientific_name)
}
\arguments{
\item{con}{A \code{seqout_connection} from \code{\link[=seqout_connect]{seqout_connect()}}.}

\item{scientific_name}{A scientific name, for example \code{"Homo sapiens"}.}
}
\value{
A tibble with \code{tax_id}, \code{scientific_name} and \code{common_name}.
}
\description{
Common name for a scientific name
}
\examples{
\dontrun{
con <- seqout_connect()
common_name(con, "Homo sapiens")
}
}
