% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R, R/zzz-aliases.R
\name{experiment_runs}
\alias{experiment_runs}
\alias{ExperimentRuns}
\title{The runs of one experiment}
\usage{
experiment_runs(accession, con = .con())

ExperimentRuns(accession, con = .con())
}
\arguments{
\item{accession}{An experiment accession (SRX, ERX, DRX, CRX, HRX).}

\item{con}{A \code{seqout_connection}. Defaults to the shared REST connection.}
}
\value{
A tibble of runs.
}
\description{
The runs of one experiment
}
