% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R, R/zzz-aliases.R
\name{project_experiments}
\alias{project_experiments}
\alias{ProjectExperiments}
\title{Get experiments for a study}
\usage{
project_experiments(con, study)

ProjectExperiments(con, study)
}
\arguments{
\item{con}{A \code{seqout_connection}.}

\item{study}{Character. Study accession (SRP/ERP/DRP/PRJ*).}
}
\value{
A tibble of experiment metadata.
}
\description{
Get experiments for a study
}
