% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R, R/zzz-aliases.R
\name{project_runs}
\alias{project_runs}
\alias{ProjectRuns}
\title{Get run download links for a study}
\usage{
project_runs(con, study)

ProjectRuns(con, study)
}
\arguments{
\item{con}{A \code{seqout_connection}.}

\item{study}{Character. Study accession.}
}
\value{
A tibble of run metadata with download links.
}
\description{
Get run download links for a study
}
