% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqout-package.R
\docType{package}
\name{seqout-package}
\alias{seqout}
\alias{seqout-package}
\title{seqout: Client for the SeqOut Genomics Database}
\description{
Query genomics study metadata from the SeqOut database (seqout.org). Provides programmatic access to metadata from NCBI GEO, SRA, ENA, and ArrayExpress. Uses DuckDB for local caching and efficient downstream analysis of query results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/saketlab/seqout}
  \item \url{https://seqout.org}
  \item Report bugs at \url{https://github.com/saketlab/seqout/issues}
}

}
\author{
\strong{Maintainer}: Saket Choudhary \email{saketlab@gmail.com} (\href{https://orcid.org/0000-0001-5202-7633}{ORCID})

}
\keyword{internal}
