% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqout-package.R
\docType{package}
\name{seqout-package}
\alias{seqout}
\alias{seqout-package}
\title{seqout: Client for Querying the 'SeqOut' Genomics Metadata Database}
\description{
Search and download genomics study metadata from 'SeqOut' \url{https://seqout.org}, which indexes NCBI GEO, NCBI SRA, EBI ENA, DDBJ (DRA and GEA), ArrayExpress and GSA.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/saketlab/seqout}
  \item \url{https://seqout.org}
  \item \url{https://seqout.org/cli/R/}
  \item Report bugs at \url{https://github.com/saketlab/seqout/issues}
}

}
\author{
\strong{Maintainer}: Saket Choudhary \email{saketc@iitb.ac.in} (\href{https://orcid.org/0000-0001-5202-7633}{ORCID})

Authors:
\itemize{
  \item Saket Choudhary \email{saketc@iitb.ac.in} (\href{https://orcid.org/0000-0001-5202-7633}{ORCID})
}

}
\keyword{internal}
