% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R, R/zzz-aliases.R
\name{seqout_close}
\alias{seqout_close}
\alias{SeqoutClose}
\title{Close a SeqOut connection}
\usage{
seqout_close(con)

SeqoutClose(con)
}
\arguments{
\item{con}{A \code{seqout_connection} returned by \code{\link[=seqout_connect]{seqout_connect()}}.}
}
\description{
Close a SeqOut connection
}
